requirements.txt
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
PyYAML>=6.0.0
Jinja2>=3.1.0
pytest>=8.0.0
